+++ src/utils/checkListCount.js
@@ -1,0 +1,6 @@
+import $ from './selector.js';
+const checkListCount = () => {
+  $('.todo-count > strong').textContent = $('.todo-list').childElementCount;
+};
+
+export default checkListCount;
